+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 2.7.0
+
+- NEW: Added a filter menu to the Personal Projects section.
+
+## 2.6.3
+
+- REVISED: Added a description to the Professional Experience and Personal Projects menu sections.
+
+## 2.6.2
+
+- REVISED: Added a confirmation dialog to the `Contact Me` button in the action menu.
+- REVISED: Added an `About` button to the action menu.
+
 ## 2.6.1
 
 - REVISED: Disabled the tilt effect on mobile devices.